feat: Adds Test stop and Validates code
Adds a lamp to indicate when the test is running, a stop test button,
and validates that all parts of the test work at least without a
subject. The code also now saves out the servo angles used for the test
which it needed to previously, but I forgot to add.

Closes #10