BREAKING CHANGE feat: Grapevine to controls angle
Changes code so that the first seven digital outputs of the Grapevine
control the pan angle of the laser and the last 5 digital outputs
control the tilt angle. The rest of the functions are updated
accordingly. For example all code to calculate angles was been deleted
and the functions that refered to the calculated angles have been
updated to use the new angle formats. Also added a Guage for test
number. Also accomplishes much of issues #13

Closes #15 issue, Closes #17 issue